Improving modularization again
